requirements.txt
faiss-cpu
sentence-transformers
anthropic
